--- quiz-game/src/components/Game/index.js
@@ -1,2 +1,0 @@
-import Game from './Game';
-export default Game;
