--- quiz-game/src/components/StartScreen/index.js
@@ -1,2 +1,0 @@
-import StartScreen from './StartScreen'
-export default StartScreen
