--- quiz-game/src/components/navigation/Route.jsx
@@ -1,3 +1,0 @@
-export function Route({children}) {
-    return children
-}
