--- quiz-game/src/context/RouterContext.jsx
@@ -1,8 +1,0 @@
-import {createContext} from "react";
-
-export const RouterContext = createContext(
-    {
-        location: "",
-        push: () => {}
-    }
-)
