--- quiz-game/src/pages/NotFound/NotFoundPage.jsx
@@ -1,9 +1,0 @@
-function NotFoundPage() {
-    return (
-        <>
-
-        </>
-    )
-}
-
-export default NotFoundPage;
