+++ quiz-game/src/pages/index.js
@@ -1,4 +1,3 @@
 export {default as MainPage} from './MainPage.jsx';
 export {default as QuizPage} from './QuizPage.jsx';
 export {default as ScorePage} from './ScorePage.jsx';
-export {default as NotFoundPage} from './NotFound/NotFoundPage.jsx';
